% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\name{render_article}
\alias{render_article}
\alias{replace_rmarkdown_engine}
\title{Render rmarkdown article}
\usage{
render_article(input, ...)

replace_rmarkdown_engine()
}
\arguments{
\item{input}{path to Rmd file}

\item{...}{passed to render}
}
\description{
Render rmarkdown article
}
