% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlmanual.R
\name{render_html_manual}
\alias{render_html_manual}
\title{Generate HTML reference manual}
\usage{
render_html_manual(package, outdir = ".")
}
\arguments{
\item{package}{name of the package}

\item{outdir}{where to put the html file}
}
\description{
Renders HTML version of the full package manual
}
