% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtools.R
\name{find_logo}
\alias{find_logo}
\alias{vignettes_base64}
\title{Build tools}
\usage{
find_logo(path = ".")

vignettes_base64(path, pkg = basename(path))
}
\arguments{
\item{path}{root directory of package}

\item{pkg}{name of the installed package}
}
\description{
Get some extra info about packages.
}
