% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtools.R, R/htmlhelp.R
\name{find_logo}
\alias{find_logo}
\alias{vignettes_base64}
\alias{commit_info_base64}
\alias{sysdeps_info}
\alias{install_sysdeps}
\alias{install_dependencies}
\alias{read_description_field}
\alias{get_ostype}
\alias{get_maintainer_info}
\alias{get_gitstats}
\alias{get_gitstats_base64}
\alias{find_readme_url}
\alias{render_readme}
\alias{render_news_files}
\alias{generate_citation_files}
\alias{get_help_metadata}
\alias{get_package_datasets}
\alias{generate_metadata_files}
\alias{needs_compilation}
\alias{maintainer_info_base64}
\alias{tweak_help_files}
\title{Build tools}
\usage{
find_logo(path, git_url, branch, subdir = "")

vignettes_base64(repo, pkg = basename(repo), subdir = "")

commit_info_base64(repo = ".")

sysdeps_info(pkg)

install_sysdeps(path = ".")

install_dependencies(path = ".")

read_description_field(fields, path = ".")

get_ostype(path = ".")

get_maintainer_info(path = ".")

get_gitstats(repo, pkgdir, url)

get_gitstats_base64(repo, pkgdir, url)

find_readme_url(url, branch = "HEAD", subdir = NULL)

render_readme(url, outdir = ".")

render_news_files(package, outdir = ".", url = NULL)

generate_citation_files(path, outdir, git_url)

get_help_metadata(package)

get_package_datasets(package)

generate_metadata_files(package, repo, subdir, outdir, pkgdir, git_url, branch)

needs_compilation(package)

maintainer_info_base64(path = ".")

tweak_help_files(path)
}
\arguments{
\item{path}{directory with html files}

\item{git_url}{of the git repository}

\item{subdir}{path within the git repo where the pkg is}

\item{repo}{path to the git repository}

\item{pkg}{name of the installed package}

\item{field}{which field from the description to show}
}
\description{
Get some extra info about packages.

Post edit HTML pages for help files
}
